requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.0
Flask==2.1.0
Flask-DotEnv==0.1.2
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
pycparser==2.21
pyOpenSSL==23.0.0
python-dotenv==0.21.1
requests==2.28.2
SQLAlchemy==1.4.41
urllib3==1.26.14
Werkzeug==2.1.0
WTForms==3.0.1
